refactor: 도메인 클래스들을 domain 패키지로 이동
