--- src/main/java/com/kakao/onboarding/precourse/albusduke/lotto/Statistics.java
@@ -1,6 +1,0 @@
-package com.kakao.onboarding.precourse.albusduke.lotto;
-
-import java.util.Map;
-
-public record Statistics(Map<Prize, Integer> counts , double ratio) {
-}
