+++ src/main/java/com/kakao/onboarding/precourse/albusduke/lotto/domain/GameResult.java
@@ -1,4 +1,4 @@
-package com.kakao.onboarding.precourse.albusduke.lotto;
+package com.kakao.onboarding.precourse.albusduke.lotto.domain;
 
 public record GameResult(int matchingCount, int bonusMatchingCount) {
 }
