+++ src/main/java/com/kakao/onboarding/precourse/albusduke/lotto/domain/LottoNumber.java
@@ -1,10 +1,12 @@
-package com.kakao.onboarding.precourse.albusduke.lotto;
+package com.kakao.onboarding.precourse.albusduke.lotto.domain;
 
 import java.util.Objects;
 
 public class LottoNumber implements Comparable<LottoNumber> {
+
     private final static int MIN_NUMBER = 1;
     private final static int MAX_NUMBER = 45;
+    private final static String NUMBER_OUT_OF_RANGE_ERR_MSG = "1 미만 또는 45 초과 숫자는 입력할 수 없습니다.";
 
     private final int number;
 
@@ -15,7 +17,7 @@ public class LottoNumber implements Comparable<LottoNumber> {
 
     private void validateRange(int number) {
         if (number > MAX_NUMBER || number < MIN_NUMBER) {
-            throw new IllegalArgumentException("1 미만 또는 45 초과 숫자는 입력할 수 없습니다.");
+            throw new IllegalArgumentException(NUMBER_OUT_OF_RANGE_ERR_MSG);
         }
     }
