+++ src/main/java/com/kakao/onboarding/precourse/albusduke/lotto/domain/Statistics.java
@@ -1,0 +1,4 @@
+package com.kakao.onboarding.precourse.albusduke.lotto.domain;
+
+public record Statistics(WinningPrizes winningPrizes, double ratio) {
+}
